// types/company.ts
export interface Company {
  companyName: string;
  contactPerson: string;

  businessEmail: string;

  phoneNumber: string;

  country: string;
  street: string;
  city: string;
  zipCode: string;

  industry: string;
  businessType: string;

  yearEstablished: string;
  employees: string;

  vatNumber: string;

  website: string;

  about: string;
}